+++ R/lccMapGenerator.R
@@ -55,7 +55,10 @@ utils::globalVariables(c(
 lccMapGenerator <- function(cohortData, pixelGroupMap,
                             lccTable = NULL,
                             deciduousCoverDiscount = 0.8418911, ## from Biomass_borealDataPrep
-                            vegLeadingProportion = 0.75,
+                            vegLeadingProportion = getOption(
+                              "NTEMS.mixedwoodProp",
+                              getOption("LandR.lccLeadingProportion", 0.75)
+                            ),
                             decidousSpp = NULL,
                             rstLCC) {
   stopifnot(