+++ R/zzz.R
@@ -1,14 +1,8 @@
 .onLoad <- function(libname, pkgname) {
   ## set options using the approach used by devtools
   opts <- options()
-  reproCachePath <- getOption("reproducible.cachePath")
-  opts.LandR <- list( #nolint
-    LandR.assertions = TRUE,
-    LandR.verbose = 1,
-    ## fetch SCANFI from the arbutus mirror when no reproducible.urlRemap is set; see
-    ## ?scanfiUrlRemap
-    LandR.scanfiMirror = TRUE
-  )
+  opts.LandR <- LandROptions()
+
   toset <- !(names(opts.LandR) %in% names(opts))
   if (any(toset)) options(opts.LandR[toset])
   .setScanfiMirror()